+++ semcore/ellipsis/src/Ellipsis.tsx
@@ -285,8 +285,9 @@ const EllipsisMiddle: React.FC<AsPropsMiddle> = (props) => {
     tooltipProps,
     children,
     advanceMode,
-    ...otherProps
+    tag,
   } = props;
+
   const resizeElement = React.useRef<HTMLDivElement>(null);
   const [dimension, setDimension] = React.useState<{ fontSize: string; symbolWidth: number }>({
     fontSize: '14',
@@ -313,7 +314,7 @@ const EllipsisMiddle: React.FC<AsPropsMiddle> = (props) => {
 
   const STail = 'span';
   const SBeginning = 'span';
-  const SContainerMiddle = props.tag || Box;
+  const SContainerMiddle = Tooltip;
   const SAdvancedModeContainerMiddle = Tooltip;
   const displayedSymbols = React.useMemo(
     () => Math.round(blockWidth / dimension.symbolWidth),
@@ -334,7 +335,7 @@ const EllipsisMiddle: React.FC<AsPropsMiddle> = (props) => {
   if (advanceMode) {
     return sstyled(styles)(
       <SAdvancedModeContainerMiddle
-        interaction={interaction}
+        interaction={tooltip ? interaction : 'none'}
         {...tooltipProps}
         {...forcedAdvancedMode}
       >
@@ -342,29 +343,20 @@ const EllipsisMiddle: React.FC<AsPropsMiddle> = (props) => {
           {children}
         </EllipsisMiddleContext.Provider>
       </SAdvancedModeContainerMiddle>,
-    ) as any;
-  }
-  if (tooltip) {
-    return sstyled(styles)(
-      <SContainerMiddle
-        interaction={interaction}
-        title={text as any}
-        ref={forkRef(ref, textRef)}
-        tag={Tooltip}
-        __excludeProps={['title']}
-        {...tooltipProps}
-      >
-        <SBeginning>{contextValue.begining}</SBeginning>
-        <STail>{contextValue.tail}</STail>
-      </SContainerMiddle>,
-    ) as any;
+    );
   }
   return sstyled(styles)(
-    <SContainerMiddle {...otherProps} ref={containerRef ?? resizeElement}>
+    <SContainerMiddle
+      interaction={tooltip ? interaction : 'none'}
+      title={text}
+      ref={forkRef(ref, textRef)}
+      tag={tag}
+      {...tooltipProps}
+    >
       <SBeginning>{contextValue.begining}</SBeginning>
       <STail>{contextValue.tail}</STail>
     </SContainerMiddle>,
-  ) as any;
+  );
 };
 
 type EllipsisContentAsProps = {