+++ stories/components/ellipsis/docs/examples/advanced_use.tsx
@@ -2,10 +2,12 @@ import { Box } from '@semcore/base-components';
 import Ellipsis from '@semcore/ellipsis';
 import React from 'react';
 
-const Demo = () => {
+import type { ExampleEllipsisProps } from '../ellipsis.stories';
+
+const Demo = (props: ExampleEllipsisProps) => {
   return (
     <Box>
-      <Ellipsis trim='middle'>
+      <Ellipsis trim={props.trim} tooltip={props.tooltip}>
         <Ellipsis.Content w={300}>
           Intergalactic, planetary, planetary, intergalactic Intergalactic, planetary, planetary,
           intergalactic Intergalactic, planetary, planetary, intergalactic Intergalactic, planetary,
@@ -16,5 +18,12 @@ const Demo = () => {
     </Box>
   );
 };
+
+export const defaultProps: ExampleEllipsisProps = {
+  trim: 'middle',
+  tooltip: true,
+};
+
+Demo.defaultProps = defaultProps;
 
 export default Demo;